+++ src/lib/firebaseConfig.js
@@ -1,0 +1,9 @@
+export const firebaseConfig = {
+    apiKey: 'AIzaSyBfP2-ChIx9Gk5O1j8j4L0WVqongsASPfI',
+    authDomain: 'social-network-bon.firebaseapp.com',
+    projectId: 'social-network-bon',
+    storageBucket: 'social-network-bon.appspot.com',
+    messagingSenderId: '927615227762',
+    appId: '1:927615227762:web:f1d0fb00d0a05d16a82feb',
+  };
+
